+++ src/react/hoc/__tests__/mutations/queries.test.tsx
@@ -3,7 +3,7 @@ import { render, wait } from '@testing-library/react';
 import gql from 'graphql-tag';
 import { DocumentNode } from 'graphql';
 
-import { ApolloClient, MutationUpdaterFn } from '../../../../core';
+import { ApolloClient, MutationUpdaterFunction, ApolloCache } from '../../../../core';
 import { ApolloProvider } from '../../../context';
 import { InMemoryCache as Cache } from '../../../../cache';
 import {
@@ -131,7 +131,12 @@ describe('graphql(mutation) query integration', () => {
       };
     }
 
-    const update: MutationUpdaterFn = (proxy, result) => {
+    const update: MutationUpdaterFunction<
+      MutationData,
+      Record<string, any>,
+      Record<string, any>,
+      ApolloCache<any>
+    > = (proxy, result) => {
       const data = JSON.parse(
         JSON.stringify(proxy.readQuery<QueryData>({ query }))
       );