+++ src/react/hoc/__tests__/mutations/recycled-queries.test.tsx
@@ -3,7 +3,7 @@ import { render, wait } from '@testing-library/react';
 import gql from 'graphql-tag';
 import { DocumentNode } from 'graphql';
 
-import { ApolloClient, MutationUpdaterFn } from '../../../../core';
+import { ApolloCache, ApolloClient, MutationUpdaterFunction } from '../../../../core';
 import { ApolloProvider } from '../../../context';
 import { InMemoryCache as Cache } from '../../../../cache';
 import { MutationFunction } from '../../../types/types';
@@ -74,7 +74,12 @@ describe('graphql(mutation) update queries', () => {
     type MutationData = typeof mutationData;
 
     let todoUpdateQueryCount = 0;
-    const update: MutationUpdaterFn = (proxy, result) => {
+    const update: MutationUpdaterFunction<
+      MutationData,
+      Record<string, any>,
+      Record<string, any>,
+      ApolloCache<any>
+    > = (proxy, result) => {
       todoUpdateQueryCount++;
       const data = JSON.parse(
         JSON.stringify(